added all

+++ set_at_placeb.c
@@ -6,7 +6,7 @@
 /*   By: nbiron <nbiron@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/29 09:49:17 by nbiron            #+#    #+#             */
-/*   Updated: 2023/10/29 09:53:46 by nbiron           ###   ########.fr       */
+/*   Updated: 2023/10/29 10:08:31 by nbiron           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static void	set_at_placeb_mini(int *stack_B)
 {
-	if (maxi(stack_B) > (ft_strlen(stack_B) / 2))
+	if ((size_t)maxi(stack_B) > (ft_strlen(stack_B) / 2))
 	{
 		while(stack_B[ft_strlen(stack_B) - 1] != maxii(stack_B))
 			rb(stack_B, ft_strlen(stack_B));
@@ -28,7 +28,7 @@ static void	set_at_placeb_mini(int *stack_B)
 
 static void	set_at_placeb_maxi(int *stack_B)
 {
-	if (maxi(stack_B) > (ft_strlen(stack_B) / 2))
+	if ((size_t)maxi(stack_B) > (ft_strlen(stack_B) / 2))
 	{
 		while(stack_B[ft_strlen(stack_B) - 1] != maxii(stack_B))
 			rb(stack_B, ft_strlen(stack_B));
@@ -55,9 +55,9 @@ void	set_at_placeb(int *stack_A, int *stack_B)
 			i++;
 		while (stack_B[j] > stack_A[0] && stack_B[j])
 			j++;
-		if (i == ft_strlen(stack_B - 1))
+		if ((size_t)i == ft_strlen(stack_B - 1))
 			set_at_placeb_maxi(stack_B);
-		else if (j == ft_strlen(stack_B - 1))
+		else if ((size_t)j == ft_strlen(stack_B - 1))
 			set_at_placeb_mini(stack_B);
 		else
 		{
